bunch of stuff

+++ ui/cpp/board_wrapper.cc
@@ -382,7 +382,7 @@ void Player::MakeMove(const v8::FunctionCallbackInfo<v8::Value>& args) {
     //     'to': {'row': int, 'col': int}}
     //   ]}
 
-    float evaluation = std::get<0>(move_res.value());
+    double evaluation = std::get<0>(move_res.value());
 
     // Convert the engine's side-to-move relative score back to an RY-relative score for the UI
     if (board->GetTurn().GetTeam() == chess::BLUE_GREEN) {
